src/record_spec: Reimplement watch_values()
This introduces a change to watch_values that makes it return a similar
Record to the one that originally spawned the Record. This creates a
form of loop where watch_values() can watch over a Record's values
until it is time to refresh the Record.

The record can be refreshed for reasons such as a timeout from the
server (due to the watch connection living too long on a timed API
gateway) or because the Record was either modified or deleted.