record_spec: Be smarter about returning Records
This fixes a potential race condition between the call of
watch_values() and the actual start of the loop. This way, it ensures
that the record that is being used is the same resource version as the
record originally passed to the function. If it is not, the function
will return the new Record, which will then be looped back over.