src/main: Sync regardless of watching
Syncing should happen regardless of using a watcher to ensure that any
extra records are deleted. This can happen due to downtime when ARES is
not watching over watched values, and a value is deleted that ARES is
unaware of deletion.

This technically does not happen in the current state but some
optimizations to watch_values (such as something that is a 1:1 relation
between a resource and a value, or an optimization to watch_values that
caches a relation between value to watched resource) could cause
resource leaks in downtime.